# Use assignment operator of std::vector to copy the entire vector in o…