I am one of the developers working on SPEC CPU's upcoming release. Minizinc is one of the many candidate benchmarks currently being considered for inclusion in the next SPEC CPU release.

As a part of this effort, it was noticed that one of the copy-constructors (for FznSpace) was a key hotspot on x86 systems. The code lines that were showing up as the hotspot were copying of elements of std::vector one-element-at-a-time in a for-loop.

Such "for-loops" were replaced with their equivalent assignment operations and that resulted in ~7% improvement in runtime of the workload (on x86).  I would expect similar gains on other architectures too.

Creating this pull request so that you can review these changes and hopefully accept them. If you have any questions or concerns (or additional suggestions) then would like to hear that as well.