@heshpdx 2024-11-12
Nice find, @rohitkprasad123 !

I can confirm the speedups reported above. Here are some times in seconds, building with gcc-14 -O3 on Ampere Altra.

threads | before | after | speedup | cmdline
-- | -- | -- | -- | --
1 | 37.42 | 34.82 | 1.075 | --two-pass gunport_problem2.mzn
4 | 78.46 | 74.01 | 1.060 | pennies.mzn pennies_7_8.dzn
4 | 161.69 | 147.36 | 1.097 | roster-sickness.mzn small-4.dzn